# cmd/api/static/openapi.yaml
openapi: 3.0.3
info:
  title: DriveBai API
  description: |
    DriveBai - Car Marketplace and Rental App API

    ## Authentication
    Most endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Error Handling
    All errors return a JSON object with an `error` field containing `code` and `message`.

    Error codes:
    - `EMAIL_TAKEN` - Email already registered
    - `INVALID_CREDENTIALS` - Wrong email or password
    - `RATE_LIMITED` - Too many requests
    - `UNAUTHORIZED` - Authentication required
    - `TOKEN_EXPIRED` - Access token expired
    - `TOKEN_INVALID` - Invalid token
    - `INVALID_ROLE` - Invalid role specified
    - `INCOMPLETE_ONBOARDING` - Onboarding requirements not met
  version: 0.2.0
  contact:
    name: DriveBai Support
    email: support@drivebai.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.drivebai.com/api/v1
    description: Production server
...
paths:
...
  /auth/token/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Exchanges a refresh token for a new access token and refresh token.
        The old refresh token is invalidated (token rotation).
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserProfile'
...
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [driver, car_owner, admin]
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
          nullable: true
        is_email_verified:
          type: boolean
        onboarding_status:
          type: string
          enum: [created, role_selected, photo_uploaded, documents_uploaded, complete]
        profile_photo_url:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: Email is required